moving createMockManifests method to testHelpers file
